requirements.txt
google-generativeai
sentence-transformers
python-dotenv
pyyaml
numpy
